DE/ppi_trading_bot/requirements.txt
ppi-client
pandas-gbq
google-auth
google-cloud-bigquery
pandas
pytz
